% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_regions.R
\name{bin_regions_nbins}
\alias{bin_regions_nbins}
\title{Bin regions using a fixed number of bins}
\usage{
bin_regions_nbins(gr, n_bins = 100, n_bins_exp = 2, id_col = "peakID")
}
\arguments{
\item{gr}{\code{GRanges} containing the regions to be divided.}

\item{n_bins}{Numeric incating the number of bins to divide the regions into
(including the extended coordinates).}

\item{n_bins_exp}{Numeric indicating the value by which to expand the regions
(\code{width(gr)*n_bins_exp}).}

\item{id_col}{Numeric or character indicating the mcol number or name of the
unique identifier for the regions.}
}
\value{
\code{GRanges} object with the peaks extended to \code{width(gr)*n_bins_exp}
and divided into a fixed number of bins.
}
\description{
Bin regions using a fixed number of bins
}
