% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wholeGenomeCor.R
\name{binGenome}
\alias{binGenome}
\title{Bin genome into windows}
\usage{
binGenome(chr_sizes, bin_size = 10000, out_type = "saf")
}
\arguments{
\item{chr_sizes}{Text files with the sizes for each chormosome, as obtain from}
}
\description{
Bin genome into windows
}
