% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTSS.R
\name{binPromoterAnnotation}
\alias{binPromoterAnnotation}
\title{Build binned promoter annotation for TSS enrichment}
\usage{
binPromoterAnnotation(
  scope = 2000,
  bin = 10,
  build = "hg38",
  out_format = "saf"
)
}
\arguments{
\item{scope}{Integer indicating how many base pairs upstream and downstream
to extend the promoter region. Returned window width will be scope*2.
Default: 2Kb.}

\item{bin}{Integer indicating the size of the bins. Default: 10 bp.}

\item{build}{Name of the build to use to extract promoter coordinates. Can be
either hg38 (default) or hg19.}
}
\value{
Either a SAF of GRanges containing binned promoter regions.
}
\description{
Build binned promoter annotation for TSS enrichment
}
