% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTSS.R
\name{binRegions}
\alias{binRegions}
\title{Bin a set of GRanges regions}
\usage{
binRegions(regions, scope = 2000, bin = 10)
}
\arguments{
\item{regions}{GRanges object containing the regions to bin. It assumes that
the first mcol contained the unique identifier of the region.}

\item{scope}{Integer indicating how many base pairs upstream and downstream
to extend the promoter region. Returned window width will be scope*2.
Default: 2Kb.}

\item{bin}{Integer indicating the size of the bins. Default: 10 bp.}
}
\value{
Binned GRanges object.
}
\description{
Bin a set of GRanges regions
}
