% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakSaturation.R
\name{.obtainOverlapMeasures}
\alias{.obtainOverlapMeasures}
\title{Obtain number of overlaps and width of the overlapping regions}
\usage{
.obtainOverlapMeasures(gr, samples)
}
\arguments{
\item{gr}{GRangesList with the peaks for each of the different samples.}

\item{samples}{Name of the samples (should be names in the GRangesList object)
to collapse together.}
}
\description{
Obtain number of overlaps and width of the overlapping regions
}
