% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTSS.R
\name{enrichmentTSS}
\alias{enrichmentTSS}
\title{Get TSS enrichment}
\usage{
enrichmentTSS(
  bam_files,
  promoter_saf,
  suffix = ".bam",
  nthreads = 5,
  paired_end = FALSE
)
}
\arguments{
\item{bam_files}{Character vector of BAM files to obtain enrichment.}

\item{promoter_saf}{SAF promoter file generated with \code{binPromoterAnnotation}.}

\item{suffix}{Suffix to remove from bam files to use as sample names. Default: ".bam".}

\item{nthreads}{Number of cores to use for the analysis.}

\item{paired_end}{Logical indicating whether the libraries are paired end or not. (Default: FALSE).}
}
\value{
Data.frame with the different promoter bins and the mean reads per sample
}
\description{
Get TSS enrichment
}
