% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conservation.R
\name{get_conservation_scores}
\alias{get_conservation_scores}
\title{Get conservation scores}
\usage{
get_conservation_scores(
  peak_file,
  id_col = 1,
  cons_score = phastCons7way.UCSC.hg38::phastCons7way.UCSC.hg38,
  bin_width = 20,
  window_width = 2000,
  n_bins = NULL,
  n_bins_exp = 2,
  merge_fun = "mean",
  random_control = TRUE,
  genome = "hg38",
  mask = NULL,
  per.chromosome = FALSE,
  summarise = TRUE
)
}
\arguments{
\item{peak_file}{Path for the peak file or GRanges object.}

\item{id_col}{Number or name of \code{mcol} that contains unique identifier for each peak. Default: 1.}

\item{cons_score}{Annotation to use for getting the conservation scores. See \code{\link[GenomicScores]{gscores}}
for details.}

\item{bin_width}{Number of base pairs for binning the peak and obtaining the conservation scores.
Ignored if the Fixed Number of Bins mode is activated (see Details). Default: 20}

\item{window_width}{Number of base pairs of a window centered in the center of the peak where to focus the
conservation analysis. Ignored if the Fixed Number of Bins mode is activated (see Details). Default = 2000}

\item{n_bins}{Integer width the number of bins each region should be divided into. Giving it a value activates
the Fixed Number of Bins mode (see Details).}

\item{n_bins_exp}{If \code{n_bins} is set, factor by which to resize the input peaks. Default: 2
(output region width = peak width * 2).}

\item{merge_fun}{Function for summarizing scores in the same region.}

\item{random_control}{Logical indicating whether to perform the same analysis using a randomized set of peaks
as a control. Default = TRUE}

\item{genome}{Character indicating the name of the genome where to randomize the regions.
See \code{\link[regioneR]{randomizeRegions}} for details.}

\item{per.chromosome}{Logical indicating if the randomization should be performed along the same chromosome as
the original set of peaks. See \code{\link[regioneR]{randomizeRegions}} for details.}

\item{summarise}{Logaical indicating whether to summarise the results, computing the mean of all peaks at a specific
position. Default: TRUE.}
}
\value{
It returns a \code{data.frame} containing the summarized phastCons conservation scores for the file
and randomized control (if \code{random_control = TRUE}) at each relative position from the peak center.
The summarization is performed by computing the mean for all peaks in that specific position. Standard deviation
is also returned.
}
\description{
Given a peak set, this function obtains the mean conservation scores in the given window and bins.
}
\details{
This function has two modes to calculate conservation scores in a set of regions: fixed window sizes (default)
or fixed number of bins.
\itemize{
\item \strong{Fixed window sizes}. Regions are resized to be \code{window_width} and are then divided into \code{bin_width} bp bins.
\item \strong{Fixed number of bins}. Regions are divided into \code{n_bins}, idependently of their size.
}
}
