% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enhancer_clusters.R
\name{get_enhancer_clusters}
\alias{get_enhancer_clusters}
\title{Obtain enhancer clusters}
\usage{
get_enhancer_clusters(
  gr,
  n_sites = 3,
  iterations = 500,
  percentile = 0.25,
  genome = "hg38"
)
}
\arguments{
\item{gr}{Character with the name of the file containing peaks or GRanges object.}

\item{n_sites}{Minimum number of sites to consider a region enhancer cluster.}

\item{iterations}{Number of iterations for obtaining distance cutoff.}

\item{percentile}{Percentile of random distances to use for stitching peaks into the same cluster. Default: 0.25.}

\item{genome}{Character indicating the name of the genome to use. Default: hg38.}
}
\value{
GRanges object containing the coordinates for the enhancer clusters
and the following mcols:
\itemize{
    \item{clustID}{Unique identifier for the enhancer cluster.}
    \item{n_sites}{Number of sites in \code{gr} that are included in that specific cluster.}
    \item{distance_cutoff}{Distance used as cutoff for including regions in the same cluster.}
}
}
\description{
Given a list of peaks, obtain clusters.
}
\details{
To define enhancer clusters, input sites are first randomized \code{iterations}
times over the \code{genome} over idividual chromosomes). Then the \code{percentile}
percentile of inter-site distances of randomized is calculated for each chromosome. Next,
clusters are defined as any group of ≥ \code{n_sites} sites in which all adjacent sites
are separated by less than the abovementioned percentile distance.
}
