% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingIndex.R
\name{get_overlaps_sign}
\alias{get_overlaps_sign}
\title{Get significant overlaps}
\usage{
get_overlaps_sign(gr_list, master, perc_overlap = 0.4)
}
\arguments{
\item{gr_list}{List of peaks with the corresponding ranking indexes, as generated by \code{\link{ranking_index}}.}

\item{master}{Master peak list.}

\item{perc_overlap}{Percentage of overlap (from 0 to 1) with a peak in a sample with the master peaklist
to consider it an overlap. Default: 0.4}
}
\description{
Get significant overlaps
}
