% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingIndex.R
\name{get_ranking_sharing_index}
\alias{get_ranking_sharing_index}
\title{Get ranking and sharing indexes}
\usage{
get_ranking_sharing_index(
  peak_files,
  bam_files,
  filename = "master_sharing_ranking_index.rda",
  perc_overlap,
  TxDb = TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene,
  nthreads = 2
)
}
\arguments{
\item{peak_files}{Character vector with the path of the peak files.}

\item{bam_files}{Character vector with the path of the bam files.}

\item{perc_overlap}{Percentage of overlap (from 0 to 1) with a peak in a sample with the master peaklist
to consider it an overlap. Default: 0.4}

\item{TxDb}{TxDb object or GRanges object with the gene annotation to use to classify peaks into
Promoters or Distal. Default: \code{TxDb.Hsapiens.UCSC.hg38.knownGene}.}

\item{nthreads}{Number of threads to use for the analysis.}
}
\description{
This function calculates sharing and ranking indexes for a set of samples and then
plots the result.
}
