% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNameFromPath.R
\name{getNameFromPath}
\alias{getNameFromPath}
\title{Get name from path}
\usage{
getNameFromPath(path, suffix = "", prefix = "")
}
\arguments{
\item{path}{Character or character vector with the paths to convert to sample names.}

\item{suffix}{Suffix to remove from the file name.}

\item{prefix}{Prefix to remove from the file name.}
}
\description{
Obtains sample names from paths.
}
