% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enhancer_clusters.R
\name{interSiteDistances}
\alias{interSiteDistances}
\title{Obtain distances between consecutive regions}
\usage{
interSiteDistances(gr)
}
\arguments{
\item{gr}{Character with the name of the file containing peaks or GRanges object.}
}
\value{
A vector of the same length as gr containing the distances to the following
region. When there is no following region (i.e. end of a chromosome), returns NA.
}
\description{
Obtain distances between consecutive regions
}
