% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccardScore.R
\name{jaccardScore}
\alias{jaccardScore}
\title{Calculate Jaccard Scores between a list of peaks}
\usage{
jaccardScore(peaks, cores = 8)
}
\arguments{
\item{peaks}{List of GRanges objects representing different sets of peaks.}

\item{cores}{Number of threads to use for obtaining overlaps.}
}
\value{
A data.frame containing the different comparisons between peaks and
the Jaccard Scores, as calculated by bedtools. As a result, the final
statistic ranges from 0.0 to 1.0, where 0.0 represents no overlap and 1.0
represent complete overlap.
}
\description{
Calculate Jaccard Scores between a list of peaks
}
