% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomPeakQual.R
\name{loadPhantomPeakQC}
\alias{loadPhantomPeakQC}
\title{Load Phantom Peak Quality Control}
\usage{
loadPhantomPeakQC(file_path, suffix = ".bam")
}
\arguments{
\item{suffix}{Suffix to remove from the bam_file to generate the sample name.}
}
\description{
Load Phantom Peak Quality Control
}
