% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalToNoiseRatios.R
\name{obtainCodingGenes}
\alias{obtainCodingGenes}
\title{Obtain protein coding genes}
\usage{
obtainCodingGenes(build = "hg38")
}
\arguments{
\item{build}{Name of the build to use to extract promoter coordinates. Can be
either hg38 (default) or hg19.}
}
\value{
Data.frame with the coordinates, strand, ensembl_gene_id and
external_gene_name of protein coding genes.#'
}
\description{
Obtains a list of protein coding genes from the human builds (hg38 or hg19).
}
