% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakSaturation.R
\name{peakSaturation}
\alias{peakSaturation}
\title{Peak saturation with increasing number of samples}
\usage{
peakSaturation(
  peakfiles,
  fc = NA,
  qval = NA,
  suffix = "_peaks.narrowPeak",
  cores = 8
)
}
\arguments{
\item{peakfiles}{Character vector of peak file paths or list of GRanges object
containing the peaks for each one of the different samples.}

\item{fc}{Numeric indicating a fold-change value to filter out peaks (from column
named "signalValue").}

\item{qval}{Numeric indicating a q-value threshold to filter out peaks (from column
named "qValue").}

\item{suffix}{Suffix to remove from the file paths to create sample names.}

\item{cores}{Number of threads to use for obtaining overlaps.}
}
\description{
Peak saturation with increasing number of samples
}
