% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phantomPeakQual.R
\name{phantomPeakQual}
\alias{phantomPeakQual}
\title{Phantom Peak Quality Control}
\usage{
phantomPeakQual(bam_file, suffix, path_phantom = "run_spp.R", nthreads = 5)
}
\arguments{
\item{bam_file}{Character with the path of the BAM file.}

\item{suffix}{Suffix to remove from the bam_file to generate the sample name.}

\item{path_phantom}{Path to the phantom peak Rscript. Default assumes it is
in your PATH.}

\item{nthreads}{Number of threads to use for the analysis.}
}
\value{
Data.frame containing the results of the QC.
}
\description{
Phantom Peak Quality Control
}
