% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingIndex.R
\name{plot_ranking_sharing_index}
\alias{plot_ranking_sharing_index}
\title{Plot sharing index vs ranking index}
\usage{
plot_ranking_sharing_index(master, plot_title)
}
\arguments{
\item{master}{Data.frame or path to the saved RData file containing the master peakset with ranking and
sharing indexes as generated with \code{\link{}}}

\item{plot_title}{Character containing the title for the final plot.}
}
\description{
This function calculates sharing and ranking indexes for a set of samples and then
plots the result.
}
