% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakSaturation.R
\name{plotPeakSaturation}
\alias{plotPeakSaturation}
\title{Plot peak saturation}
\usage{
plotPeakSaturation(res, comp = "peaks", title, xlab = "# of samples")
}
\arguments{
\item{res}{Results of the \code{peakSaturation} function.}

\item{comp}{Name of the comparison to plot. It can be either "peaks" (default),
which will plot the number of overlapping peaks or "bp" which will overlap the
number of overlapping bp.}

\item{title}{Character with the title of the plot}

\item{xlab}{Label for the x axis.}
}
\description{
Plot peak saturation
}
