% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingIndex.R
\name{ranking_index}
\alias{ranking_index}
\title{Obtain ranking index}
\usage{
ranking_index(peak_file, bam_file, nthreads = 2)
}
\arguments{
\item{peak_file}{Character containing the path for the peak file.}

\item{bam_file}{Character containing the path for the bam file.}

\item{nthreads}{Number of threads to use for this analysis.}
}
\description{
Given a set of peak files, obtains the ranking index for each one of them.
}
