% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingIndex.R
\name{sharing_index}
\alias{sharing_index}
\title{Obtain sharing index}
\usage{
sharing_index(gr_list, perc_overlap = 0.4)
}
\arguments{
\item{gr_list}{List of peaks with the corresponding ranking indexes, as generated by \code{\link{ranking_index}}.}

\item{perc_overlap}{Percentage of overlap (from 0 to 1) with a peak in a sample with the master peaklist
to consider it an overlap. Default: 0.4}
}
\description{
Obtain sharing index
}
