% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wholeGenomeCor.R
\name{wholeGenomeCor}
\alias{wholeGenomeCor}
\title{Whole-genome correlation between samples}
\usage{
wholeGenomeCor(
  bam_files,
  chr_sizes,
  suffix = ".bam",
  bin_size = 10000,
  nthreads = 3,
  paired_end = FALSE
)
}
\arguments{
\item{bam_files}{Character vector of bam file paths.}

\item{chr_sizes}{Text files with the sizes for each chormosome, as obtain from}

\item{suffix}{Suffix to remove from bam files to use as sample names. Default: ".bam".}

\item{nthreads}{Number of cores to use for the analysis.}

\item{paired_end}{Logical indicating whether the libraries are paired end or not. (Default: FALSE).}
}
\description{
Whole-genome correlation between samples
}
