% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalToNoiseRatios.R
\name{obtainCodingGenes}
\alias{obtainCodingGenes}
\title{Obtain protein coding genes}
\usage{
obtainCodingGenes(build = "hg38")
}
\arguments{
\item{build}{Name of the build to use to extract promoter coordinates. Can be
either hg38 (default) or hg19.}
}
\description{
Obtain protein coding genes
}
